requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3

# Geospatial & Remote Sensing
earthengine-api>=0.1.350
geemap>=0.30.0
rasterio>=1.3.0
geopandas>=0.14.0

# Image Processing
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0

# Jupyter (for notebook usage)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# PDF Processing (optional)
PyMuPDF>=1.23.0
